Lay in initial Cloud Build configs for generating demo apps on Cloud Run
This is experimental for now.

There are two configs in this changeset, one for mozorg and a very similar
one for pocket. They both trigger builds of containers, uploads them to
Artifact Registry in us-central1 and deoploys them to Cloud Run in either
Mozorg mode or Pocket mode, respectively.

This is not the entire puzzle, though. We also will need to...

1) Set up corresponding Services with the expected names, which are taken
from the branch names: mozorg-demo-1, mozorg-demo-2....5, pocket-demo-1,
pocket-demo-2 (We'll just have two pocket demos for now)

2) Set up custom domains to those (now-known) services

As for setting custom env vars - that's still to be solved.