Update demo config to use separate env file(s)
We can't support multiple env files, so developers can't have a core env
and then add to a 'demo.env' additional file, unfortunately.

However, they can edit/add to these new Mozorg- or Pocket-specific env files.
There may be an occaisonal merge conflict in these files, but that should
be very rare.

Secrets are kept secret and need to be provisioned separately. At the moment
we're going with a somewhat verbose and numerous set of secrets, but maybe
we can pull all this into a single YAML file of secrets, like we do with k8s.